pyproject.toml
[build-system]
requires = ['poetry-core>=1.2.0', 'wheel',]
build-backend = 'poetry.core.masonry.api'

[tool.poetry]
name = 'trestlebot'
version = '0.1.0'
description = "trestle-bot assists users in leveraging Compliance-Trestle in automated workflows of for OSCAL formatted compliance content management."

authors = ["Jennifer Power <jpower@redhat.com>",]

include = ['LICENSE']
license = 'Apache-2.0'
readme = 'README.md'

repository = 'https://github.com/RedHatProductSecurity/trestle-bot'

[tool.poetry.scripts]
trestle-bot = "trestlebot.cli:run"

[tool.poetry.dependencies]
python = '^3.8.1'
gitpython = "^3.1.35"
compliance-trestle = {git = "https://github.com/IBM/compliance-trestle.git", rev = "01a5a3358b7f1c35d98ea5ab32d51bbc107ff5a0"}
github3-py = "^4.0.1"
python-gitlab = "^3.15.0"
ruamel-yaml = "^0.17.32"
pydantic = "1.10.12"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
isort = "^5.12.0"
safety = "^2.3.5"
bandit = "^1.7.5"
flake8-print = "^5.0.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    'tests/*',
]

[tool.pytest.ini_options]
minversion = '6.0'
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    'tests',
]

[tool.mypy]

plugins = [
  "pydantic.mypy"
]

follow_imports = "skip"
warn_redundant_casts = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "github3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ruamel"
ignore_missing_imports = true
